Model is now functionally finished, with the exception of planTrip()
planTrip() is like the main of Model, and am kinda waiting to see
exactly what View and Presenter want Model to do, before I implement it.
But, the coolest part I think is getLegDistance() method, which will
give us a number representation (in km) of the distance between two
locations.  In tripCo the main currently is set up to test some
distances, check it out.